Fix signed integer overflows in surface provider cost estimate
Errors observed in unit tests:
image/AnimationSurfaceProvider.cpp:110:39, signed integer overflow: 590076 * 49429 cannot be represented in type 'int$
image/DecodedSurfaceProvider.cpp:125:21, signed integer overflow: 65534 * 65531 cannot be represented in type 'int'
image/SurfaceCache.cpp:77:22, signed integer overflow: 256 * 17895696 cannot be represented in type 'int'
image/SurfaceFilters.h:92:50, signed integer overflow: 60000 * 60000 cannot be represented in type 'int'

+++ image/AnimationSurfaceProvider.h
@@ -44,7 +44,7 @@ public:
   DrawableSurface Surface() override { return DrawableSurface(WrapNotNull(this)); }
 
   bool IsFinished() const override;
-  size_t LogicalSizeInBytes() const override;
+  double LogicalSizeInBytes() const override;
   void AddSizeOfExcludingThis(MallocSizeOf aMallocSizeOf,
                               size_t& aHeapSizeOut,
                               size_t& aNonHeapSizeOut,
